requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
Pillow==8.3.1
psycopg2-binary==2.9.1
pyparsing==2.4.7
pyTelegramBotAPI==3.8.2
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.6
